requirements.txt
tabulate
requests
python-decouple
pytest